Move all batchManager methods to the correct file
